INFUND-6923 WIP PublicContentItemServiceImpl
Former-commit-id: 4a4c8c17b3baa36b6e34ea663df54b29478a99b0

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/publiccontent/transactional/PublicContentItemService.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface PublicContentItemService {
 
     @NotSecured("Visible on for everyone even not logged in users")
-    ServiceResult<PublicContentItemPageResource> findFilteredItems(Optional<Long> innovationAreaId, Optional<String> searchString, Optional<Long> pageNumber, Optional<Long> pageSize);
+    ServiceResult<PublicContentItemPageResource> findFilteredItems(Optional<Long> innovationAreaId, Optional<String> searchString, Optional<Integer> pageNumber, Optional<Integer> pageSize);
 
     @NotSecured("Visible on for everyone even not logged in users")
     ServiceResult<PublicContentItemResource> byCompetitionId(Long id);
